Introduces Greetings Repo. in Java. Reviews main README.md
